requirements.txt
# AI Orchestration
langchain
langchain-core
langchain-community
crewai
python-dotenv
msgspec
pydantic

# LLM Providers (Targeted)
langchain-google-genai
langchain-groq

# Vector Database (Local ChromaDB)
chromadb
langchain-chroma
sentence-transformers

# Research Tools (Optional but recommended)
duckduckgo-search
tavily-python

# Frontend & Auth
streamlit
streamlit-option-menu
supabase

# General Utilities
rich
pandas
altair
plotly
ipywidgets
ipykernel
